impl. ft_strlen, as well as test

+++ tests/test_read.c
@@ -1,10 +1,10 @@
-# include <stdio.h>
+#include <stdio.h>
 #include <libasm.h>
 #include <unistd.h>
 
 int test_read(int fd, void *buf, size_t count)
 {
-	int	ret;
+	ssize_t	ret;
 
 	printf("Using ft_read\n");
 	ret = ft_read(fd, buf, count);
@@ -14,7 +14,7 @@ int test_read(int fd, void *buf, size_t count)
 		return (ret);
 	}
 	((char *)buf)[ret]=0;
-	printf("read chars: %d %s\n", ret, (char *)buf);
+	printf("read chars: %zd %s\n", ret, (char *)buf);
 
 	printf("Using read\n");
 	ret = read(fd, buf, count);
@@ -24,6 +24,6 @@ int test_read(int fd, void *buf, size_t count)
 		return (ret);
 	}
 	((char *)buf)[ret]=0;
-	printf("read chars: %d %s\n", ret, (char *)buf);
+	printf("read chars: %zd %s\n", ret, (char *)buf);
 	return (0);
 }
