+++ tests/test_write.c
@@ -1,10 +1,10 @@
-# include <stdio.h>
+#include <stdio.h>
 #include <libasm.h>
 #include <unistd.h>
 
 int test_write(int fd, const void *buf, size_t count)
 {
-	int	ret;
+	ssize_t	ret;
 
 	printf("Using ft_write\n");
 	ret = ft_write(fd, buf, count);
@@ -13,7 +13,7 @@ int test_write(int fd, const void *buf, size_t count)
 		perror("Error");
 		return (ret);
 	}
-	printf("written chars: %d\n", ret);
+	printf("written chars: %zd\n", ret);
 
 	printf("Using write\n");
 	ret = write(fd, buf, count);
@@ -22,6 +22,6 @@ int test_write(int fd, const void *buf, size_t count)
 		perror("Error");
 		return (ret);
 	}
-	printf("written chars: %d\n", ret);
+	printf("written chars: %zd\n", ret);
 	return (0);
 }
